backend/ml_services/requirements.txt
tensorflow-cpu>=2.18.0
opencv-python>=4.5.0
numpy>=1.19.0
pillow
scikit-learn>=0.24.0
gdown>=4.0.0
flask
flask-cors
gunicorn
